requirements.txt
# Core libraries for deep learning and image processing
torch
torchvision
opencv-python
Pillow
numpy